Major, (possible breaking) Changes
Shared PatchFileResolver — no duplicated URL logic.
All four commands support --patches with URLs.
`--prerelease` on all commands
`--temporary-files-path` with shared `morphe-temporary-files/` default
Cache reuse across commands
ListCompatibleVersions upgraded from positional arg to `--patches` flag
