+++ src/main/kotlin/app/morphe/cli/command/ListPatchesCommand.kt
@@ -11,8 +11,11 @@ package app.morphe.cli.command
 import app.morphe.patcher.patch.Package
 import app.morphe.patcher.patch.Patch
 import app.morphe.patcher.patch.loadPatchesFromJar
+import picocli.CommandLine
 import picocli.CommandLine.Command
 import picocli.CommandLine.Option
+import picocli.CommandLine.Spec
+import picocli.CommandLine.Model.CommandSpec
 import picocli.CommandLine.Help.Visibility.ALWAYS
 import java.io.File
 import java.util.logging.Logger
@@ -32,7 +35,28 @@ internal object ListPatchesCommand : Runnable {
         arity = "1..*",
         required = true
     )
-    private lateinit var patchFiles: Set<File>
+    @Suppress("unused")
+    private fun setPatchesFile(patchesFiles: Set<File>) {
+        patchesFiles.firstOrNull {
+            !it.exists() &&
+                    !it.toString().startsWith("http:/") &&
+                    !it.toString().startsWith("https:/")
+        }?.let {
+            throw CommandLine.ParameterException(
+                spec.commandLine(),
+                "${it.name} can't be found"
+            )
+        }
+        this.patchesFiles = patchesFiles
+    }
+    private var patchesFiles = emptySet<File>()
+
+    @Option(
+        names = ["--prerelease"],
+        description = ["Fetch the latest dev pre-release instead of the stable main release from the repo provided in --patches."],
+        showDefaultValue = ALWAYS,
+    )
+    private var prerelease: Boolean = false
 
     @Option(
         names = ["--out"],
@@ -46,6 +70,12 @@ internal object ListPatchesCommand : Runnable {
         showDefaultValue = ALWAYS,
     )
     private var withDescriptions: Boolean = true
+
+    @Option(
+        names = ["-t", "--temporary-files-path"],
+        description = ["Path to store temporary files."],
+    )
+    private var temporaryFilesPath: File? = null
 
     @Option(
         names = ["-p", "--with-packages"],
@@ -87,6 +117,9 @@ internal object ListPatchesCommand : Runnable {
         description = ["Filter patches by package name."],
     )
     private var packageName: String? = null
+
+    @Spec
+    private lateinit var spec: CommandSpec
 
     override fun run() {
         fun Package.buildString(): String {
@@ -157,7 +190,23 @@ internal object ListPatchesCommand : Runnable {
                 compatiblePackageName == name
             } ?: withUniversalPatches
 
-        val patches = loadPatchesFromJar(patchFiles).withIndex().toList()
+
+        val temporaryFilesPath = temporaryFilesPath ?: File("").absoluteFile.resolve("morphe-temporary-files")
+
+        try {
+            patchesFiles = PatchFileResolver.resolve(
+                patchesFiles,
+                prerelease,
+                temporaryFilesPath
+            )
+        } catch (e: IllegalArgumentException) {
+            throw CommandLine.ParameterException(
+                spec.commandLine(),
+                e.message ?: "Failed to resolve patch URL"
+            )
+        }
+
+        val patches = loadPatchesFromJar(patchesFiles).withIndex().toList()
 
         val filtered = packageName?.let {
             patches.filter { (_, patch) ->
@@ -172,7 +221,7 @@ internal object ListPatchesCommand : Runnable {
         val finalOutput = filtered.joinToString("\n\n") {it.buildString()}
 
         if (filtered.isEmpty()) {
-            logger.warning("No compatible patches found in: $patchFiles")
+            logger.warning("No compatible patches found in: $patchesFiles")
         } else {
             if (outputFile == null) {
                 logger.info(finalOutput)
