+++ src/main/kotlin/app/morphe/cli/command/OptionsCommand.kt
@@ -8,6 +8,8 @@ import app.morphe.patcher.patch.loadPatchesFromJar
 import kotlinx.serialization.json.Json
 import picocli.CommandLine
 import picocli.CommandLine.Command
+import picocli.CommandLine.Option
+import picocli.CommandLine.Help.Visibility.ALWAYS
 import picocli.CommandLine.Model.CommandSpec
 import picocli.CommandLine.Spec
 import java.io.File
@@ -28,29 +30,49 @@ internal object OptionsCommand : Callable<Int> {
     @Spec
     private lateinit var spec: CommandSpec
 
-    @CommandLine.Option(
+    @Option(
         names = ["-p", "--patches"],
         description = ["One or more paths to MPP files."],
         required = true,
     )
     @Suppress("unused")
     private fun setPatchesFile(patchesFiles: Set<File>) {
-        patchesFiles.firstOrNull { !it.exists() }?.let {
-            throw CommandLine.ParameterException(spec.commandLine(), "${it.name} can't be found")
+        patchesFiles.firstOrNull {
+            !it.exists() &&
+            !it.toString().startsWith("http:/") &&
+            !it.toString().startsWith("https:/")
+        }?.let {
+            throw CommandLine.ParameterException(
+                spec.commandLine(),
+                "${it.name} can't be found"
+            )
         }
         this.patchesFiles = patchesFiles
     }
 
     private var patchesFiles = emptySet<File>()
 
-    @CommandLine.Option(
+    @Option(
         names = ["-o", "--out"],
         description = ["Path to the output JSON file."],
         required = true,
     )
     private lateinit var outputFile: File
 
-    @CommandLine.Option(
+    @Option(
+        names = ["--prerelease"],
+        description = ["Fetch the latest dev pre-release instead of the stable main release from the repo provided in --patches."],
+        showDefaultValue = ALWAYS,
+    )
+    private var prerelease: Boolean = false
+
+    @Option(
+        names = ["-t", "--temporary-files-path"],
+        description = ["Path to store temporary files."],
+    )
+    private var temporaryFilesPath: File? = null
+
+    @Option(
         names = ["-f", "--filter-package-name"],
         description = ["Filter patches by compatible package name."],
     )
@@ -59,6 +81,21 @@ internal object OptionsCommand : Callable<Int> {
     private val json = Json { prettyPrint = true }
 
     override fun call(): Int {
+        val temporaryFilesPath = temporaryFilesPath ?: File("").absoluteFile.resolve("morphe-temporary-files")
+
+        try {
+            patchesFiles = PatchFileResolver.resolve(
+                patchesFiles,
+                prerelease,
+                temporaryFilesPath
+            )
+        } catch (e: IllegalArgumentException) {
+            throw CommandLine.ParameterException(
+                spec.commandLine(),
+                e.message ?: "Failed to resolve patch URL"
+            )
+        }
+
         return try {
             logger.info("Loading patches")
